Lots added here
Added Junit tests for customer, bike, and location classes. there is an issue with trying to add providers and bikes when mocksystem is initialised so currently it is added in the customers class. working on getting it to be done in the junit tests. Overall tweaks to work with current variable types